moved file encoding to general export options
