+++ impexp-config/build.gradle
@@ -3,7 +3,7 @@ plugins {
 }
 
 dependencies {
-    api 'org.citygml4j:citygml4j:2.11.3-b20210126-1315'
+    api 'org.citygml4j:citygml4j:2.11.3-b20210126.2019'
 }
 
 bintray {
