+++ impexp-config/src/main/java/org/citydb/config/project/exporter/GeneralOptions.java
@@ -1,14 +1,24 @@
 package org.citydb.config.project.exporter;
 
 import javax.xml.bind.annotation.XmlType;
+import java.nio.charset.StandardCharsets;
 
 @XmlType(name="GeneralExportOptionsType", propOrder={})
 public class GeneralOptions {
+    private String fileEncoding;
     private OutputFormat compressedOutputFormat = OutputFormat.CITYGML;
     private ExportEnvelope envelope;
 
     public GeneralOptions() {
         envelope = new ExportEnvelope();
+    }
+
+    public String getFileEncoding() {
+        return fileEncoding != null ? fileEncoding : StandardCharsets.UTF_8.name();
+    }
+
+    public void setFileEncoding(String fileEncoding) {
+        this.fileEncoding = fileEncoding;
     }
 
     public OutputFormat getCompressedOutputFormat() {
